// src/model/rest/transport/PreDefMoneyflowTransport.ts
export type PreDefMoneyflowTransport = {
  id: number;
  userid: number;
  amount: number;
  capitalsourceid: number;
  capitalsourcecomment: string;
  contractpartnerid: number;
  contractpartnername: string;
  comment: string;
  createdate: string;
  onceAMonth: number;
  lastUsed: string;
  postingaccountid: number;
  postingaccountname: string;
};